requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
azure-common==1.1.23
azure-functions==1.0.0b5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
DateTime==4.3
freezegun==0.3.12
httplib2==0.13.0
icalendar==4.0.3
ics==0.4
idna==2.8
importlib-metadata==0.18
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
zope.interface==4.6.0